rename providers file with capitalization
